requirements.txt
asgiref==3.6.0
Django==4.2
django-filter==22.1
djangorestframework==3.15.2
djangorestframework-simplejwt==4.6.0
pytz==2023.3
sqlparse==0.4.4
psycopg2-binary==2.9.10
drf-spectacular==0.28.0
drf-standardized-errors==0.14.1
